# tms/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ScyllaX-TMS API in Go using Fiber framework",
        "title": "ScyllaX-TMS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/mobile/pickup/skip": {
            "post": {
                "description": "save skip pickup to db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skip pickup"
                ],
                "summary": "skip pickup",
                "parameters": [
                    {
                        "description": "skip pickup",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SkipPickUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
...
        "/mobile/skip": {
            "post": {
                "description": "update skip_at to db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visit"
                ],
                "summary": "skip",
                "parameters": [
                    {
                        "description": "skip",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SkipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
...
        "/web/shipment-report/reason": {
            "get": {
                "description": "Return reason for dropdown.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web(shipment_report)-dropdown"
                ],
                "summary": "get reason dropdown.",
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.ReasonDropdown"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.JsonBadRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "email": "email is required",
                        "username": "username is required"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "BAD REQUEST"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
...
        "entity.JsonInternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "errors": {
                    "type": "string",
                    "example": "error database or third party"
                },
                "status": {
                    "type": "string",
                    "example": "INTERNAL SERVER ERROR"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
        "entity.JsonNotFound": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "errors": {
                    "type": "string",
                    "example": "record not found"
                },
                "status": {
                    "type": "string",
                    "example": "NOT FOUND"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
        "entity.JsonSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
...
        "entity.ReasonDropdown": {
            "type": "object",
            "properties": {
                "reason_name": {
                    "type": "string",
                    "example": "reject"
                }
            }
        },
...
        "entity.SkipPickUpBody": {
            "type": "object",
            "required": [
                "id",
                "reason_id",
                "reason_name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1218
                },
                "reason_id": {
                    "type": "integer",
                    "example": 1
                },
                "reason_name": {
                    "type": "string",
                    "example": "Reject"
                }
            }
        },
        "entity.SkipPickUpRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SkipPickUpBody"
                    }
                }
            }
        },
        "entity.SkipRequest": {
            "type": "object",
            "required": [
                "current_time",
                "in_outlet",
                "outlet_id",
                "shipment_no",
                "skip_reason"
            ],
            "properties": {
                "current_time": {
                    "type": "integer"
                },
                "in_outlet": {
                    "type": "boolean"
                },
                "outlet_id": {
                    "type": "integer"
                },
                "shipment_no": {
                    "type": "string"
                },
                "skip_reason": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}